requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
langchain-ollama